==> code/tnm_stage/1-Compile_Dataset.ipynb <==
import os
import pandas as pd

# Data loading

data_dir = "../../data"

## Corpus

df_corpus = pd.read_csv(
    os.path.join(data_dir, "corpus", "TCGA_Reports.csv"),
    sep=',',
    header=0
)

df_corpus.shape

df_corpus.head()

Extract TCGA patient IDs:

df_corpus["patient_id"] = df_corpus["patient_filename"].apply(lambda x: x.split('.')[0])

assert not df_corpus["patient_filename"].duplicated().any()

assert not df_corpus["patient_id"].duplicated().any()

df_corpus = df_corpus[["patient_id", "text"]]

df_corpus.index = df_corpus["patient_id"].values

df_corpus.head()

## TNM staging

### Tumor size (T)

df_tnm_t = pd.read_csv(
    os.path.join(data_dir, "tnm_stage", "TCGA_T14_patients.csv"),
    sep=',',
    header=0
)

df_tnm_t.shape

df_tnm_t.head()

df_tnm_t.rename(
    columns={"case_submitter_id": "patient_id"},
    inplace=True
)

assert not df_tnm_t["patient_id"].duplicated().any()

df_tnm_t.index = df_tnm_t["patient_id"].values

df_tnm_t = df_tnm_t[["patient_id", "ajcc_pathologic_t"]]

df_tnm_t.head()

len(set(df_tnm_t["ajcc_pathologic_t"].values))

df_tnm_t["ajcc_pathologic_t"].value_counts(dropna=False)

#### Preprocess labels

def label_t_stage(t_value):
    # Extract the first digit after 'T'
    if t_value.startswith('T') and len(t_value) > 1:
        first_char = t_value[1]
        if first_char in ['1', '2', '3', '4']:
            return t_value[:2]
    return None

df_tnm_t["t_label"] = df_tnm_t["ajcc_pathologic_t"].apply(label_t_stage)

df_tnm_t.shape

df_tnm_t = df_tnm_t[df_tnm_t["t_label"].notna()]

df_tnm_t.shape

df_tnm_t["t_label"].value_counts(dropna=False)

### Regional lymph node involvement (N)

df_tnm_n = pd.read_csv(
    os.path.join(data_dir, "tnm_stage", "TCGA_N03_patients.csv"),
    sep=',',
    header=0
)

df_tnm_n.shape

df_tnm_n.head()

df_tnm_n.rename(
    columns={"case_submitter_id": "patient_id"},
    inplace=True
)

assert not df_tnm_n["patient_id"].duplicated().any()

df_tnm_n.index = df_tnm_n["patient_id"].values

df_tnm_n = df_tnm_n[["patient_id", "ajcc_pathologic_n"]]

df_tnm_n.head()

len(set(df_tnm_n["ajcc_pathologic_n"].values))

df_tnm_n["ajcc_pathologic_n"].value_counts(dropna=False)

#### Preprocess labels

def label_n_stage(n_value):
    # Extract the first digit after 'N'
    if n_value.startswith('N') and len(n_value) > 1:
        first_char = n_value[1]
        if first_char in ['0', '1', '2', '3']:
            return n_value[:2]
    return None

df_tnm_n["n_label"] = df_tnm_n["ajcc_pathologic_n"].apply(label_n_stage)

df_tnm_n.shape

df_tnm_n = df_tnm_n[df_tnm_n["n_label"].notna()]

df_tnm_n.shape

df_tnm_n["n_label"].value_counts(dropna=False)

### Distant metastasis (M)

df_tnm_m = pd.read_csv(
    os.path.join(data_dir, "tnm_stage", "TCGA_M01_patients.csv"),
    sep=',',
    header=0
)

df_tnm_m.shape

df_tnm_m.head()

df_tnm_m.rename(
    columns={"case_submitter_id": "patient_id"},
    inplace=True
)

assert not df_tnm_m["patient_id"].duplicated().any()

df_tnm_m.index = df_tnm_m["patient_id"].values

df_tnm_m = df_tnm_m[["patient_id", "ajcc_pathologic_m"]]

df_tnm_m.head()

len(set(df_tnm_m["ajcc_pathologic_m"].values))

df_tnm_m["ajcc_pathologic_m"].value_counts(dropna=False)

#### Preprocess labels

def label_m_stage(m_value):
    # Extract the first digit after 'M'
    if m_value.startswith('M') and len(m_value) > 1:
        first_char = m_value[1]
        if first_char in ['0', '1']:
            return m_value[:2]
    return None

df_tnm_m["m_label"] = df_tnm_m["ajcc_pathologic_m"].apply(label_m_stage)

df_tnm_m.shape

df_tnm_m = df_tnm_m[df_tnm_m["m_label"].notna()]

df_tnm_m.shape

df_tnm_m["m_label"].value_counts(dropna=False)

# Data merging

We combine all datasets into a single dataframe.

df_tnm_t.shape, df_tnm_n.shape, df_tnm_m.shape

We select patients annotated with T, N and M labels:

arr_patient_id = sorted(
    set(df_tnm_t["patient_id"]).intersection(
        set(df_tnm_n["patient_id"])
    ).intersection(
        set(df_tnm_m["patient_id"])
    )
)

len(arr_patient_id)

assert pd.Series(arr_patient_id).isin(df_corpus["patient_id"]).all()

df_corpus_tnm = df_corpus.loc[arr_patient_id].copy()

df_corpus_tnm["t_label"] = df_tnm_t.loc[arr_patient_id]["t_label"].values
df_corpus_tnm["n_label"] = df_tnm_n.loc[arr_patient_id]["n_label"].values
df_corpus_tnm["m_label"] = df_tnm_m.loc[arr_patient_id]["m_label"].values

df_corpus_tnm.shape

df_corpus_tnm.head()

df_corpus_tnm.to_csv(
    os.path.join(data_dir, "tnm_stage", "tcga_reports_tnm_stage.csv"),
    sep=',',
    index=False,
    header=True
)
