==> code/tnm_stage/3-BoW_LR.ipynb <==
import os
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import plotly.express as px

import nltk
nltk.download('punkt_tab')  # This line is only needed the first time you run the code
nltk.download('stopwords') # This line is only needed the first time you run the code

from nltk.tokenize import word_tokenize

from sklearn.feature_extraction.text import CountVectorizer, TfidfTransformer
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline

from sklearn.metrics import accuracy_score, classification_report, f1_score, recall_score, precision_score

# Utils

arr_stopwords = nltk.corpus.stopwords.words('english')

def tokenizer(text):
    """
    Tokenizes the input text (lowercased) using NLTK's word_tokenize function and removes
    stopwords, punctuation, tokens containing non-alphabetic characters, and tokens with length 1.
    
    Args:
        text (str): The input text to tokenize.
        
    Returns:
        list: A list of tokens.
    """
    arr_tokens = word_tokenize(text.lower())
    # Remove tokens with length 1
    arr_tokens = [token for token in arr_tokens if len(token) > 1]
    # Remove stopwords
    arr_tokens = [token for token in arr_tokens if token not in arr_stopwords]
    # Remove punctuation and tokens containing non-alphabetic characters
    arr_tokens = [token for token in arr_tokens if token.isalpha()]
    return arr_tokens

def calculate_performance(arr_gs, arr_preds, arr_labels, col_label, df_data, df_train_data):
    dict_perf = {
        'precision': precision_score(
            y_true=arr_gs,
            y_pred=arr_preds,
            average=None,
            labels=arr_labels,
            zero_division=0.0
        ),
        'recall': recall_score(
            y_true=arr_gs,
            y_pred=arr_preds,
            average=None,
            labels=arr_labels,
            zero_division=0.0
        ),
        'f1': f1_score(
            y_true=arr_gs,
            y_pred=arr_preds,
            average=None,
            labels=arr_labels,
            zero_division=0.0
        )
    }
    arr_res = []
    for i in range(len(arr_labels)):
        label = arr_labels[i]
        p, r, f1 = dict_perf['precision'][i], dict_perf['recall'][i], dict_perf['f1'][i]
        n_train = len(df_train_data[df_train_data[col_label] == label])
        n_eval = len(df_data[df_data[col_label] == label])
        arr_res.append({
            "label": label,
            "precision": p,
            "recall": r,
            "f1": f1,
            "n_train": n_train,
            "n_val": n_eval
        })
    df_perf = pd.DataFrame(arr_res)

    df_perf.sort_values(
        by=["f1", "recall", "precision"],
        ascending=True
    )

    return df_perf

# Data loading

data_dir = "../../data/tnm_stage"

## Training

df_train = pd.read_csv(
    os.path.join(data_dir, "train_tcga_reports_tnm_stage.csv"),
    sep=',',
    header=0
)

df_train.shape

df_train.head()

arr_train_corpus = df_train["text"].values.tolist()

## Validation

df_val = pd.read_csv(
    os.path.join(data_dir, "val_tcga_reports_tnm_stage.csv"),
    sep=',',
    header=0
)

df_val.shape

df_val.head()

arr_val_corpus = df_val["text"].values.tolist()

# Multiclass (single task) approach

arr_train_labels = df_train["tnm_label"].values.tolist()

arr_val_labels = df_val["tnm_label"].values.tolist()

## Logistic Regression

lr_pipeline = Pipeline([
    (
        "bow", 
        CountVectorizer(
            tokenizer=tokenizer,
            token_pattern=None,
            lowercase=False,
            stop_words=None
        )
    ),
    (
        "tf-idf",
        TfidfTransformer(
            use_idf=True
        )
    ),
    (
        "classifier",
        LogisticRegression(
            random_state=0
        )
    )
])

lr_pipeline.fit(arr_train_corpus, arr_train_labels)

### Evaluation

#### Train

arr_train_pred = lr_pipeline.predict(arr_train_corpus)

accuracy_score(
    y_true=arr_train_labels,
    y_pred=arr_train_pred
)

print(classification_report(
    y_true=arr_train_labels,
    y_pred=arr_train_pred
))

#### Val

arr_val_pred = lr_pipeline.predict(arr_val_corpus)

accuracy_score(
    y_true=arr_val_labels,
    y_pred=arr_val_pred
)

calculate_performance(
    arr_gs=arr_val_labels,
    arr_preds=arr_val_pred,
    arr_labels=lr_pipeline.classes_,
    col_label="tnm_label",
    df_data=df_val,
    df_train_data=df_train
)

# Multi-task approach

# T subtask

arr_t_train_labels = df_train["t_label"].values.tolist()

arr_t_val_labels = df_val["t_label"].values.tolist()

## Logistic Regression

lr_t_pipeline = Pipeline([
    (
        "bow", 
        CountVectorizer(
            tokenizer=tokenizer,
            token_pattern=None,
            lowercase=False,
            stop_words=None
        )
    ),
    (
        "tf-idf",
        TfidfTransformer(
            use_idf=True
        )
    ),
    (
        "classifier",
        LogisticRegression(
            random_state=0
        )
    )
])

lr_t_pipeline.fit(arr_train_corpus, arr_t_train_labels)

### Evaluation

#### Train

arr_t_train_pred = lr_t_pipeline.predict(arr_train_corpus)

accuracy_score(
    y_true=arr_t_train_labels,
    y_pred=arr_t_train_pred
)

print(classification_report(
    y_true=arr_t_train_labels,
    y_pred=arr_t_train_pred
))

#### Val

arr_t_val_pred = lr_t_pipeline.predict(arr_val_corpus)

accuracy_score(
    y_true=arr_t_val_labels,
    y_pred=arr_t_val_pred
)

calculate_performance(
    arr_gs=arr_t_val_labels,
    arr_preds=arr_t_val_pred,
    arr_labels=lr_t_pipeline.classes_,
    col_label="t_label",
    df_data=df_val,
    df_train_data=df_train
)

# N subtask

arr_n_train_labels = df_train["n_label"].values.tolist()

arr_n_val_labels = df_val["n_label"].values.tolist()

## Logistic Regression

lr_n_pipeline = Pipeline([
    (
        "bow", 
        CountVectorizer(
            tokenizer=tokenizer,
            token_pattern=None,
            lowercase=False,
            stop_words=None
        )
    ),
    (
        "tf-idf",
        TfidfTransformer(
            use_idf=True
        )
    ),
    (
        "classifier",
        LogisticRegression(
            random_state=0
        )
    )
])

lr_n_pipeline.fit(arr_train_corpus, arr_n_train_labels)

### Evaluation

#### Train

arr_n_train_pred = lr_n_pipeline.predict(arr_train_corpus)

accuracy_score(
    y_true=arr_n_train_labels,
    y_pred=arr_n_train_pred
)

print(classification_report(
    y_true=arr_n_train_labels,
    y_pred=arr_n_train_pred
))

#### Val

arr_n_val_pred = lr_n_pipeline.predict(arr_val_corpus)

accuracy_score(
    y_true=arr_n_val_labels,
    y_pred=arr_n_val_pred
)

calculate_performance(
    arr_gs=arr_n_val_labels,
    arr_preds=arr_n_val_pred,
    arr_labels=lr_n_pipeline.classes_,
    col_label="n_label",
    df_data=df_val,
    df_train_data=df_train
)

# M subtask

arr_m_train_labels = df_train["m_label"].values.tolist()

arr_m_val_labels = df_val["m_label"].values.tolist()

## Logistic Regression

lr_m_pipeline = Pipeline([
    (
        "bow", 
        CountVectorizer(
            tokenizer=tokenizer,
            token_pattern=None,
            lowercase=False,
            stop_words=None
        )
    ),
    (
        "tf-idf",
        TfidfTransformer(
            use_idf=True
        )
    ),
    (
        "classifier",
        LogisticRegression(
            random_state=0
        )
    )
])

lr_m_pipeline.fit(arr_train_corpus, arr_m_train_labels)

### Evaluation

#### Train

arr_m_train_pred = lr_m_pipeline.predict(arr_train_corpus)

accuracy_score(
    y_true=arr_m_train_labels,
    y_pred=arr_m_train_pred
)

print(classification_report(
    y_true=arr_m_train_labels,
    y_pred=arr_m_train_pred
))

#### Val

arr_m_val_pred = lr_m_pipeline.predict(arr_val_corpus)

accuracy_score(
    y_true=arr_m_val_labels,
    y_pred=arr_m_val_pred
)

calculate_performance(
    arr_gs=arr_m_val_labels,
    arr_preds=arr_m_val_pred,
    arr_labels=lr_m_pipeline.classes_,
    col_label="m_label",
    df_data=df_val,
    df_train_data=df_train
)

# TNM staging
We finally compare the performance of the Multi-task vs the Multiclass approach for the TNM staging classification problem:

arr_tnm_train_pred = arr_t_train_pred + arr_n_train_pred + arr_m_train_pred
arr_tnm_val_pred = arr_t_val_pred + arr_n_val_pred + arr_m_val_pred

# Train

# Single Multiclass approach
accuracy_score(
    y_true=arr_train_labels,
    y_pred=arr_train_pred
)

# Multitask approach
accuracy_score(
    y_true=arr_train_labels,
    y_pred=arr_tnm_train_pred
)

# Val

# Single Multiclass approach
accuracy_score(
    y_true=arr_val_labels,
    y_pred=arr_val_pred
)

# Multitask approach
accuracy_score(
    y_true=arr_val_labels,
    y_pred=arr_tnm_val_pred
)
