==> code/tnm_stage/5-BERT_Eval.ipynb <==
import os
import pickle
import pandas as pd

import sys
sys.path.append('..')
import utils

from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import accuracy_score

# Constants and arguments
data_dir = "../../data/tnm_stage"
out_preds_path = "./model_preds"
model_name = "Clinical-BigBird"

# Data loading

t_label_enc = LabelEncoder()
n_label_enc = LabelEncoder()
m_label_enc = LabelEncoder()

## Training

df_train = pd.read_csv(os.path.join(data_dir, "train_tcga_reports_tnm_stage.csv"))

df_train.shape

t_label_enc = t_label_enc.fit(df_train['t_label'])
n_label_enc = n_label_enc.fit(df_train['n_label'])
m_label_enc = m_label_enc.fit(df_train['m_label'])

## Validation

df_val = pd.read_csv(os.path.join(data_dir, "val_tcga_reports_tnm_stage.csv"))

df_val.shape

## Test

df_test = pd.read_csv(os.path.join(data_dir, "test_tcga_reports_tnm_stage.csv"))

df_test.shape

# Multi-task approach

## T subtask

with open(
    os.path.join(out_preds_path, f"t_label_{model_name}_test_preds.pkl"),
    'rb'
) as file:
    t_test_preds = pickle.load(file)

arr_t_test_label_preds = t_label_enc.inverse_transform(t_test_preds.argmax(axis=1))

accuracy_score(
    y_true=df_test[f't_label'].values,
    y_pred=arr_t_test_label_preds
)

utils.calculate_performance(
    arr_gs=df_test[f't_label'].values,
    arr_preds=arr_t_test_label_preds,
    arr_labels=t_label_enc.classes_,
    col_label=f"t_label",
    df_data=df_test,
    df_train_data=df_train
)

## N subtask

with open(
    os.path.join(out_preds_path, f"n_label_{model_name}_test_preds.pkl"),
    'rb'
) as file:
    n_test_preds = pickle.load(file)

arr_n_test_label_preds = n_label_enc.inverse_transform(n_test_preds.argmax(axis=1))

accuracy_score(
    y_true=df_test[f'n_label'].values,
    y_pred=arr_n_test_label_preds
)

utils.calculate_performance(
    arr_gs=df_test[f'n_label'].values,
    arr_preds=arr_n_test_label_preds,
    arr_labels=n_label_enc.classes_,
    col_label=f"n_label",
    df_data=df_test,
    df_train_data=df_train
)

## M subtask

with open(
    os.path.join(out_preds_path, f"m_label_{model_name}_test_preds.pkl"),
    'rb'
) as file:
    m_test_preds = pickle.load(file)

arr_m_test_label_preds = m_label_enc.inverse_transform(m_test_preds.argmax(axis=1))

accuracy_score(
    y_true=df_test[f'm_label'].values,
    y_pred=arr_m_test_label_preds
)

utils.calculate_performance(
    arr_gs=df_test[f'm_label'].values,
    arr_preds=arr_m_test_label_preds,
    arr_labels=m_label_enc.classes_,
    col_label=f"m_label",
    df_data=df_test,
    df_train_data=df_train
)
