==> code/tnm_stage/6-LLM_TNM.ipynb <==
This notebook uses an LLM to perform TNM staging classification.

import os
import pickle
from datetime import timedelta
import numpy as np
import pandas as pd
import time
import json

from enum import Enum
from pydantic import BaseModel
from pydantic.generics import GenericModel
from typing import TypeVar, Generic, Any

import sys
sys.path.append('..')
import utils
import llm_utils

import torch
from transformers import AutoTokenizer
from vllm import LLM, SamplingParams
from vllm.sampling_params import GuidedDecodingParams
from sklearn.metrics import accuracy_score

os.environ['VLLM_WORKER_MULTIPROC_METHOD'] = 'spawn'

# Model
model_size = "8B"
model_name = f"/common/gonzalezg7lab/Projects/LLMs/externalModels/Meta-Llama-3.1-{model_size}-Instruct"

data_dir = "../data/mimic_corpus/test.csv"
prompt_path = "./prompts/"
out_path = f"./model_preds/tnm_label_llama_{model_size}_prompt_{{prompt}}_{{dataset}}_preds.csv"
prompt_path = "./prompts"

max_input_len = 90000
max_output_len = 1024

cuda_gpu_id = "0"

data_dir = "../../data/tnm_stage"

dict_tnm_labels = {
    "t_label": ["T1", "T2", "T3", "T4"],
    "n_label": ["N0", "N1", "N2", "N3"],
    "m_label": ["M0", "M1"],
}

if cuda_gpu_id != "-1":
    os.environ["CUDA_VISIBLE_DEVICES"] = cuda_gpu_id
_ = torch.device('cuda')
torch.backends.cuda.matmul.allow_tf32 = True
assert torch.cuda.is_available()
print("Number of GPUs available:", torch.cuda.device_count())

# Data loading

## Training

df_train = pd.read_csv(os.path.join(data_dir, "train_tcga_reports_tnm_stage.csv"))

df_train.shape

assert not df_train['patient_id'].duplicated().any()

df_train.index = df_train['patient_id'].values

## Validation

df_val = pd.read_csv(os.path.join(data_dir, "val_tcga_reports_tnm_stage.csv"))

df_val.shape

assert not df_val['patient_id'].duplicated().any()

df_val.index = df_val['patient_id'].values

## Test

df_test = pd.read_csv(os.path.join(data_dir, "test_tcga_reports_tnm_stage.csv"))

df_test.shape

assert not df_test['patient_id'].duplicated().any()

df_test.index = df_test['patient_id'].values

# Prompts

prompt_name = "tnm_zs"

user_msg_template = """Pathology report: "{text}"
Output in JSON format: """

# Structured output JSON format
class T_Value(str, Enum):
    t1 = "T1"
    t2 = "T2"
    t3 = "T3"
    t4 = "T4"


class N_Value(str, Enum):
    n0 = "N0"
    n1 = "N1"
    n2 = "N2"
    n3 = "N3"


class M_Value(str, Enum):
    m0 = "M0"
    m1 = "M1"


# Generic type variable
E = TypeVar("E", T_Value, N_Value, M_Value)


# Generic TNM item model
class TNM_Item(GenericModel, Generic[E]):
    explanation: Any
    label: E


# Final TNM format model
class TNM_Format(BaseModel):
    T: TNM_Item[T_Value]
    N: TNM_Item[N_Value]
    M: TNM_Item[M_Value]

arr_prompt = [
    {
        'role': 'system',
        'content': """You are a medical expert tasked with extracting cancer stage information from pathology reports. Based on the text provided, classify the tumor stage (T), lymph node involvement (N), and metastasis (M) according to standard TNM classification guidelines. Provide both a classification label and a brief natural language explanation for each category. If information is not explicit, infer the most likely value based on the context.

- T can take on 4 integer values: 1 if tumor is 2cm or less across, 2 if tumor is more than 2cm but not more than 5cm across, 3 if tumor is more than 5cm across, or 4 if tumor of any size growing into the chest wall or skin. Return no text beyond the integer value.
- N can take on 4 integer values: 0 if there are no cancer cells in any nearby nodes or only small clusters of cancer cells less than 0.2 mm across, 1 if there are cancer cells in 1 to 3 lymph nodes, 2 if there are 4 to 9 lymph nodes in the armpit and at least one is larger than 2 mm, or 3 if there are cancer cells in 10 or more lymph nodes in the armpit and at least one is larger than 2 mm."
- M can take on 2 integer values: 0 if there is no distant metastasis, or 1 if there is distant metastasis (i.e. cancer that has spread from the original (primary) tumor to distant organs or distant lymph nodes)."


Return your answer strictly in the following JSON format:

{
  "T": {
    "explanation": [Your reasoning for T classification],
    "label": [T classification label]
  },
  "N": {
    "explanation": [Your reasoning for N classification],
    "label": [N classification label]
  },
  "M": {
    "explanation": [Your reasoning for M classification],
    "label": [M classification label]
  }
}"""
    }
]

# Save the prompt
with open(os.path.join(prompt_path, f"{prompt_name}.json"), 'w') as f:
    json.dump(arr_prompt, f)

# Model loading

tokenizer = AutoTokenizer.from_pretrained(model_name)

We first analyze the token length of each document in the corpus:

arr_input_prompt = llm_utils.create_prompt(
    df_eval=pd.DataFrame({'text': df_train['text'].to_list() + df_val['text'].to_list() + df_test['text'].to_list()}),
    func_format_user=llm_utils.func_format_user,
    messages=arr_prompt,
    user_template=user_msg_template
)

arr_tok_prompt = [
    tokenizer.apply_chat_template(
        prompt,
        add_generation_prompt=True,
        tokenize=True
    )
    for prompt in arr_input_prompt
]

arr_tok_len = pd.Series([len(seq) for seq in arr_tok_prompt])
print(arr_tok_len.describe())

print(pd.DataFrame({
    "abs": (arr_tok_len <= max_input_len).value_counts(normalize=False),
    "rel": (arr_tok_len <= max_input_len).value_counts(normalize=True)
}))
print()

All documents fit into the model.

number_gpus = torch.cuda.device_count()

print("Loading ", model_name)
start_time = time.time()
model = LLM(
    model=model_name,
    tensor_parallel_size=number_gpus,
    dtype=torch.bfloat16,
    gpu_memory_utilization=.90,
    max_model_len=max_input_len
)
end_time = time.time()
print("Model loaded!")
print("Total loading time:", str(timedelta(seconds=end_time - start_time)))
print()

sampling_params = SamplingParams(
    temperature=.6,
    top_p=.9,
    max_tokens=max_output_len,
    seed=0,
    detokenize=True,
    guided_decoding=GuidedDecodingParams(
        json=TNM_Format.model_json_schema()
    )
)

# Evaluation

## Validation

arr_val_input_prompt = llm_utils.create_prompt(
    df_eval=df_val,
    func_format_user=llm_utils.func_format_user,
    messages=arr_prompt,
    user_template=user_msg_template
)

print("Number of texts to predict:", len(arr_val_input_prompt))

start_time = time.time()
arr_val_text_pred = llm_utils.eval_prompt(
    arr_input_prompt=arr_val_input_prompt,
    tokenizer=tokenizer,
    model=model,
    sampling_params=sampling_params
)
end_time = time.time()

print("Total inference time:", str(timedelta(seconds=end_time - start_time)))

df_val_pred = llm_utils.extract_preds(
    df_eval=df_val,
    arr_preds=arr_val_text_pred
)

### T

tnm_label = "t_label"

accuracy_score(
    y_true=df_val[tnm_label].values,
    y_pred=df_val_pred[tnm_label].values
)

utils.calculate_performance(
    arr_gs=df_val[tnm_label].values,
    arr_preds=df_val_pred[tnm_label].values,
    arr_labels=dict_tnm_labels[tnm_label],
    col_label=tnm_label,
    df_data=df_val,
    df_train_data=df_train
)

### N

tnm_label = "n_label"

accuracy_score(
    y_true=df_val[tnm_label].values,
    y_pred=df_val_pred[tnm_label].values
)

utils.calculate_performance(
    arr_gs=df_val[tnm_label].values,
    arr_preds=df_val_pred[tnm_label].values,
    arr_labels=dict_tnm_labels[tnm_label],
    col_label=tnm_label,
    df_data=df_val,
    df_train_data=df_train
)

### M

tnm_label = "m_label"

accuracy_score(
    y_true=df_val[tnm_label].values,
    y_pred=df_val_pred[tnm_label].values
)

utils.calculate_performance(
    arr_gs=df_val[tnm_label].values,
    arr_preds=df_val_pred[tnm_label].values,
    arr_labels=dict_tnm_labels[tnm_label],
    col_label=tnm_label,
    df_data=df_val,
    df_train_data=df_train
)

We save the model predictions:

df_val_pred.to_csv(
    out_path.format(
        prompt=prompt_name,
        dataset="val"
    ),
    index=False
)

## Test

arr_test_input_prompt = llm_utils.create_prompt(
    df_eval=df_test,
    func_format_user=llm_utils.func_format_user,
    messages=arr_prompt,
    user_template=user_msg_template
)

print("Number of texts to predict:", len(arr_test_input_prompt))

start_time = time.time()
arr_test_text_pred = llm_utils.eval_prompt(
    arr_input_prompt=arr_test_input_prompt,
    tokenizer=tokenizer,
    model=model,
    sampling_params=sampling_params
)
end_time = time.time()

print("Total inference time:", str(timedelta(seconds=end_time - start_time)))

df_test_pred = llm_utils.extract_preds(
    df_eval=df_test,
    arr_preds=arr_test_text_pred
)

### T

tnm_label = "t_label"

accuracy_score(
    y_true=df_test[tnm_label].values,
    y_pred=df_test_pred[tnm_label].values
)

utils.calculate_performance(
    arr_gs=df_test[tnm_label].values,
    arr_preds=df_test_pred[tnm_label].values,
    arr_labels=dict_tnm_labels[tnm_label],
    col_label=tnm_label,
    df_data=df_test,
    df_train_data=df_train
)

### N

tnm_label = "n_label"

accuracy_score(
    y_true=df_test[tnm_label].values,
    y_pred=df_test_pred[tnm_label].values
)

utils.calculate_performance(
    arr_gs=df_test[tnm_label].values,
    arr_preds=df_test_pred[tnm_label].values,
    arr_labels=dict_tnm_labels[tnm_label],
    col_label=tnm_label,
    df_data=df_test,
    df_train_data=df_train
)

### M

tnm_label = "m_label"

accuracy_score(
    y_true=df_test[tnm_label].values,
    y_pred=df_test_pred[tnm_label].values
)

utils.calculate_performance(
    arr_gs=df_test[tnm_label].values,
    arr_preds=df_test_pred[tnm_label].values,
    arr_labels=dict_tnm_labels[tnm_label],
    col_label=tnm_label,
    df_data=df_test,
    df_train_data=df_train
)

We save the model predictions:

df_test_pred.to_csv(
    out_path.format(
        prompt=prompt_name,
        dataset="test"
    ),
    index=False
)
