==> code/tnm_stage/7-LLM_Eval.ipynb <==
import os
import pickle
import pandas as pd

import sys
sys.path.append('..')
import utils

from sklearn.metrics import accuracy_score

# Constants and arguments
data_dir = "../../data/tnm_stage"
out_preds_path = "./model_preds"
model_name = "llama_8B"
prompt_name = "tnm_zs"
out_path = os.path.join(out_preds_path, f"tnm_label_{model_name}_prompt_{prompt_name}_{{dataset}}_preds.csv")

# Data loading

## Training

df_train = pd.read_csv(os.path.join(data_dir, "train_tcga_reports_tnm_stage.csv"))

df_train.shape

assert not df_train['patient_id'].duplicated().any()
df_train.index = df_train['patient_id'].values

## Test

df_test = pd.read_csv(os.path.join(data_dir, "test_tcga_reports_tnm_stage.csv"))

df_test.shape

assert not df_test['patient_id'].duplicated().any()
df_test.index = df_test['patient_id'].values

## Model preds

### Test

df_test_pred = pd.read_csv(out_path.format(dataset="test"))

df_test_pred.shape

assert not df_test_pred['patient_id'].duplicated().any()
df_test_pred.index = df_test_pred['patient_id'].values

# T subtask

accuracy_score(
    y_true=df_test['t_label'].values,
    y_pred=df_test_pred.loc[df_test['patient_id'].values, 't_label'].values
)

utils.calculate_performance(
    arr_gs=df_test['t_label'].values,
    arr_preds=df_test_pred.loc[df_test['patient_id'].values, 't_label'].values,
    arr_labels=sorted(df_test['t_label'].unique()),
    col_label=f"t_label",
    df_data=df_test,
    df_train_data=df_train
)

# N subtask

accuracy_score(
    y_true=df_test[f'n_label'].values,
    y_pred=df_test_pred.loc[df_test['patient_id'].values, 'n_label'].values
)

utils.calculate_performance(
    arr_gs=df_test['n_label'].values,
    arr_preds=df_test_pred.loc[df_test['patient_id'].values, 'n_label'].values,
    arr_labels=sorted(df_test['n_label'].unique()),
    col_label=f"n_label",
    df_data=df_test,
    df_train_data=df_train
)

# M subtask

accuracy_score(
    y_true=df_test['m_label'].values,
    y_pred=df_test_pred.loc[df_test['patient_id'].values, 'm_label'].values
)

utils.calculate_performance(
    arr_gs=df_test[f'm_label'].values,
    arr_preds=df_test_pred.loc[df_test['patient_id'].values, 'm_label'].values,
    arr_labels=sorted(df_test['m_label'].unique()),
    col_label=f"m_label",
    df_data=df_test,
    df_train_data=df_train
)
